requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# JavaScript Rendering (for WAF-protected sites)
playwright>=1.40.0

# Excel Export
openpyxl>=3.1.0

# Telegram Notifications
python-telegram-bot>=20.0

# Scheduling
schedule>=1.2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0